requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
altair==4.1.0
appdirs==1.4.4
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.1.0
astor==0.8.1
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
clikit==0.6.2
cmdstanpy==0.9.5
convertdate==2.3.2
crashtest==0.3.1
cycler==0.11.0
Cython==0.29.24
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
ephem==4.1
fbprophet==0.7.1
fonttools==4.28.1
frozenlist==1.2.0
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
hijri-converter==2.2.2
holidays==0.11.3.1
httpstan==4.6.1
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
korean-lunar-calendar==0.2.1
llvmlite==0.37.0
LunarCalendar==0.0.9
lxml==4.6.4
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
multitasking==0.0.10
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.6
numba==0.54.1
numpy==1.20.3
packaging==21.3
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
py-lets-be-rational==1.0.1
py-vollib==1.0.1
py-vollib-vectorized==0.1.1
pyarrow==6.0.1
pycodestyle==2.8.0
pycparser==2.21
pydeck==0.7.1
Pygments==2.10.0
pylev==1.4.0
PyMeeus==0.5.11
Pympler==0.9
pyparsing==3.0.6
pyrsistent==0.18.0
pysimdjson==3.2.0
pystan==2.19.1.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.26.0
scipy==1.7.2
Send2Trash==1.8.0
setuptools-git==1.2
setuptools-scm==6.3.2
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.27
streamlit==1.2.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webargs==8.0.1
webencodings==0.5.1
widgetsnbextension==3.5.2
yarl==1.7.2
yfinance==0.1.67
zipp==3.6.0
